# docs/contracts/v1/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchAcceptedResponse": {
        "description": "Acknowledges a batch, applied or recognized as a replay.",
        "properties": {
          "applied": {
            "description": "Observations stored by this call. Zero for a replay.",
            "title": "Applied",
            "type": "integer"
          },
          "batch_id": {
            "format": "uuid",
            "title": "Batch Id",
            "type": "string"
          },
          "duplicate": {
            "description": "True when this batch_id had already been applied.",
            "title": "Duplicate",
            "type": "boolean"
          },
          "edges_written": {
            "default": 0,
            "title": "Edges Written",
            "type": "integer"
          },
          "principals_written": {
            "default": 0,
            "title": "Principals Written",
            "type": "integer"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "servers_written": {
            "default": 0,
            "title": "Servers Written",
            "type": "integer"
          },
          "share_aces_written": {
            "default": 0,
            "title": "Share Aces Written",
            "type": "integer"
          },
          "shares_written": {
            "default": 0,
            "title": "Shares Written",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "batch_id",
          "applied",
          "duplicate"
        ],
        "title": "BatchAcceptedResponse",
        "type": "object"
      },
...
      "CollectorKind": {
        "description": "Which collector produced an observation.",
        "enum": [
          "active_directory",
          "smb",
          "ntfs",
          "local_groups"
        ],
        "title": "CollectorKind",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScanRunListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScanRunSummaryView"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "ScanRunListResponse",
        "type": "object"
      },
      "ScanRunStart": {
        "additionalProperties": false,
        "description": "Opens a scan run.\n\nRe-posting an identical start is a no-op that returns the existing run: the collector\nowns ``run_id``, so a retry after a timeout cannot create a second run.",
        "properties": {
          "incremental": {
            "default": false,
            "title": "Incremental",
            "type": "boolean"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "default": "1.3",
            "title": "Schema Version",
            "type": "string"
          },
          "scopes": {
            "items": {
              "$ref": "#/components/schemas/Scope"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Scopes",
            "type": "array"
          },
          "source": {
            "$ref": "#/components/schemas/SourceDescriptor"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "source",
          "started_at",
          "scopes"
        ],
        "title": "ScanRunStart",
        "type": "object"
      },
      "ScanRunStartedResponse": {
        "description": "Acknowledges a start envelope.",
        "properties": {
          "created": {
            "description": "True when this call opened the run; false when it replayed a start.",
            "title": "Created",
            "type": "boolean"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status",
          "created"
        ],
        "title": "ScanRunStartedResponse",
        "type": "object"
      },
...
      "ScanRunView": {
        "description": "Everything known about one run.",
        "properties": {
          "batch_count_received": {
            "title": "Batch Count Received",
            "type": "integer"
          },
          "batch_count_reported": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Count Reported"
          },
          "collector": {
            "title": "Collector",
            "type": "string"
          },
          "collector_host": {
            "title": "Collector Host",
            "type": "string"
          },
          "collector_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collector Version"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "declared_scopes": {
            "items": {
              "$ref": "#/components/schemas/ScopeView"
            },
            "title": "Declared Scopes",
            "type": "array"
          },
          "downgrade_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Downgrade Reason"
          },
          "error_count": {
            "title": "Error Count",
            "type": "integer"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/CollectorErrorView"
            },
            "title": "Errors",
            "type": "array"
          },
          "incremental": {
            "title": "Incremental",
            "type": "boolean"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "observation_count_applied": {
            "title": "Observation Count Applied",
            "type": "integer"
          },
          "observation_count_reported": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation Count Reported"
          },
          "reconciled_scopes": {
            "items": {
              "$ref": "#/components/schemas/ScopeView"
            },
            "title": "Reconciled Scopes",
            "type": "array"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "run_id",
          "status",
          "incremental",
          "started_at",
          "completed_at",
          "collector",
          "collector_host",
          "method",
          "collector_version",
          "target",
          "batch_count_reported",
          "batch_count_received",
          "observation_count_reported",
          "observation_count_applied",
          "error_count",
          "notes",
          "downgrade_reason",
          "declared_scopes",
          "reconciled_scopes",
          "errors"
        ],
        "title": "ScanRunView",
        "type": "object"
      },
      "ScanStatus": {
        "description": "Lifecycle of a scan run.",
        "enum": [
          "pending",
          "running",
          "succeeded",
          "partial",
          "failed",
          "canceled"
        ],
        "title": "ScanStatus",
        "type": "string"
      },
      "Scope": {
        "additionalProperties": false,
        "description": "A reconciliation scope: the boundary inside which absence may be inferred.",
        "properties": {
          "key": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ScopeKind"
          }
        },
        "required": [
          "kind",
          "key"
        ],
        "title": "Scope",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Windows-domain share-access auditing and governance",
    "title": "ADG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scan-runs": {
      "get": {
        "operationId": "list_scan_runs_api_v1_scan_runs_get",
        "parameters": [
          {
            "description": "Restrict to one collector kind.",
            "in": "query",
            "name": "collector",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CollectorKind"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to one collector kind.",
              "title": "Collector"
            }
          },
          {
            "description": "Restrict to runs in one lifecycle state.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ScanStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to runs in one lifecycle state.",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 500,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "description": "Opaque cursor from a previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page.",
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "The cursor is not one this endpoint issued."
          }
        },
        "summary": "Scan runs, newest first",
        "tags": [
          "ingestion"
        ]
      },
      "post": {
        "operationId": "start_scan_run_api_v1_scan_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ADG-Collector-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Adg-Collector-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRunStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The same start was replayed; the existing run is returned."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRunStartedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The run id exists but describes a different run."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Begin a scan run",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/scan-runs/{run_id}": {
      "get": {
        "operationId": "get_scan_run_api_v1_scan_runs__run_id__get",
        "parameters": [
          {
            "description": "The run id the collector generated and reuses on every retry.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "The run id the collector generated and reuses on every retry.",
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No such run."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect a scan run",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/scan-runs/{run_id}/batches": {
      "post": {
        "operationId": "submit_batch_api_v1_scan_runs__run_id__batches_post",
        "parameters": [
          {
            "description": "The run id the collector generated and reuses on every retry.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "The run id the collector generated and reuses on every retry.",
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-ADG-Collector-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Adg-Collector-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No such run; send the start envelope first."
          },
          "409": {
            "description": "The run is already completed and accepts no more observations."
          },
          "422": {
            "description": "The payload is invalid, or carries a kind this phase cannot store."
          }
        },
        "summary": "Send a batch of observations",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/scan-runs/{run_id}/completion": {
      "post": {
        "operationId": "complete_scan_run_api_v1_scan_runs__run_id__completion_post",
        "parameters": [
          {
            "description": "The run id the collector generated and reuses on every retry.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "The run id the collector generated and reuses on every retry.",
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-ADG-Collector-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Adg-Collector-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRunCompletion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRunCompletedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No such run."
          },
          "409": {
            "description": "The run was already completed with a different outcome, or the completion reconciles a scope the run never declared."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close a scan run",
        "tags": [
          "ingestion"
        ]
      }
    },
...
  }
}